requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
chromadb>=0.4.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
tqdm>=4.65.0
